Rename normalization2 module to normalization
