+++ external/fv3fit/fv3fit/train_microphysics.py
@@ -29,7 +29,7 @@ from fv3fit import set_random_seed
 from fv3fit._shared import put_dir
 from fv3fit._shared.config import OptimizerConfig
 from fv3fit._shared.hyperparameters import Hyperparameters
-from fv3fit.emulation.layers.normalization2 import MeanMethod, StdDevMethod
+from fv3fit.emulation.layers.normalization import MeanMethod, StdDevMethod
 from fv3fit.keras._models.shared.pure_keras import PureKerasDictPredictor
 from fv3fit.keras.jacobian import compute_jacobians, nondimensionalize_jacobians
 
@@ -65,7 +65,7 @@ from fv3fit.emulation.zhao_carr.models import PrecpdModelConfig
 from fv3fit.emulation.zhao_carr.filters import HighAntarctic
 from fv3fit.emulation.flux import TendencyToFlux, MoistStaticEnergyTransform
 
-from fv3fit.emulation.layers.normalization2 import standard_deviation_all_features
+from fv3fit.emulation.layers.normalization import standard_deviation_all_features
 from fv3fit.wandb import (
     WandBConfig,
     store_model_artifact,
